add Create Post endpoint

+++ src/main/java/com/fethibey/social/repository/DbInitializer.java
@@ -27,9 +27,9 @@ public class DbInitializer implements CommandLineRunner {
             var post = new Post();
             post.setTitle(faker.book().title());
             post.setContent(faker.lorem().paragraph());
-            post.setViews(random.nextLong(0,100));
-            post.setLikes(random.nextLong(0,50));
-            post.setShares(random.nextLong(0,10));
+            post.setViews(random.nextInt(0,100));
+            post.setLikes(random.nextInt(0,50));
+            post.setShares(random.nextInt(0,10));
             post.setPublicationDate(new Date(faker.date().past(10, TimeUnit.DAYS).toInstant().toEpochMilli()));
             postRepository.save(post);
         }
